implementing small fry .. untestet
