+++ graphlearn/utils/draw.py
@@ -3,7 +3,7 @@ from eden.util.display import draw_graph
 import networkx as nx
 import graphlearn.graphtools as graphtools
 from collections import defaultdict
-
+from graphlearn.utils import calc_stats_from_grammar
 import logging
 logger = logging.getLogger(__name__)
 
@@ -315,20 +315,7 @@ def draw_graph_row(graph,
         # plt.show()
 
 
-def calc_stats_from_grammar(grammar):
-    count_corehashes = defaultdict(int)
-    count_interfacehashes = defaultdict(int)
-    corecounter = defaultdict(int)
-    intercounter = defaultdict(int)
-    for ih in grammar.keys():
-        for ch in grammar[ih].keys():
-            # go over all the combos
-            count_corehashes[ch]+=1
-            count_interfacehashes[ih]+=1
-            count= grammar[ih][ch].count
-            corecounter[ch]+=count
-            intercounter[ih]+=count
-    return count_corehashes,count_interfacehashes,corecounter,intercounter
+
 
 def contract_edges(original_graph):
     """
